modify addQR dialog and add location checkbox functionality.
