+++ app/src/main/java/com/example/qrhunter/fragments/ScannerFragment.java
@@ -47,7 +47,7 @@ import java.util.Map;
 /**
  * This class handles QRCode scanning using the zxing-android-embedded library
  */
-public class ScannerFragment extends Fragment {
+public class ScannerFragment extends Fragment{
     private final String TAG = "Scanner Fragment";
     private onCameraClose listener;
 
@@ -115,34 +115,34 @@ public class ScannerFragment extends Fragment {
         if (result.getContents() == null) { //User either went back or qrcode contents are empty
             goToWallet();
         } else {
-            GeoPoint geoPoint = getLocation();
-            System.out.println(geoPoint);
+//            GeoPoint geoPoint = getLocation();
+//            System.out.println(geoPoint);
             // Generate hash from qrcode contents
             String hash = Hashing.sha256().hashString(result.getContents(), StandardCharsets.UTF_8).toString();
 
-            QrCodeScoreGenerator scoreGenerator = new QrCodeScoreGenerator();
-            int score = scoreGenerator.score_algorithm(hash);
-
-            QrCodeNameGenerator nameGenerator = new QrCodeNameGenerator();
-            String codeName = nameGenerator.createQRName(hash);
-
-            simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm");
-            String date = simpleDateFormat.format(new Date());
-            System.out.println(date);
-
-            // Put QRCode into database
-            Map<String, Object> QRInfo = new HashMap<>();
-            QRInfo.put("name", codeName);
-            QRInfo.put("date", date);
-            QRInfo.put("hash", hash);
-            QRInfo.put("owner", owner);
-            QRInfo.put("location", geoPoint);
-            QRInfo.put("score", score);
-            CollectionReference CodeList = db.collection("CodeList");
-            CodeList.add(QRInfo);
+//            QrCodeScoreGenerator scoreGenerator = new QrCodeScoreGenerator();
+//            int score = scoreGenerator.score_algorithm(hash);
+//
+//            QrCodeNameGenerator nameGenerator = new QrCodeNameGenerator();
+//            String codeName = nameGenerator.createQRName(hash);
+//
+//            simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd HH:mm");
+//            String date = simpleDateFormat.format(new Date());
+//            System.out.println(date);
+//
+//            // Put QRCode into database
+//            Map<String, Object> QRInfo = new HashMap<>();
+//            QRInfo.put("name", codeName);
+//            QRInfo.put("date", date);
+//            QRInfo.put("hash", hash);
+//            QRInfo.put("owner", owner);
+//            QRInfo.put("location", geoPoint);
+//            QRInfo.put("score", score);
+//            CollectionReference CodeList = db.collection("CodeList");
+//            CodeList.add(QRInfo);
 
             // Open dialog showing user the qrcode they just scanned
-            QrCodeOnAddDialog qrAddDialog = new QrCodeOnAddDialog(hash, getActivity());
+            QrCodeOnAddDialog qrAddDialog = new QrCodeOnAddDialog(hash, getActivity(),owner);
             qrAddDialog.show(getParentFragmentManager(),"Test" );
             goToWallet();
         }
@@ -152,52 +152,52 @@ public class ScannerFragment extends Fragment {
     /**
      * This method checks if app has permission for location access and ask if doesn't.
      */
-    private void askPermission() {
+    public void askPermission() {
         if (ActivityCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_FINE_LOCATION) != PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_COARSE_LOCATION) != PackageManager.PERMISSION_GRANTED) {
             ActivityCompat.requestPermissions(getActivity(), new String[]{Manifest.permission.ACCESS_FINE_LOCATION, Manifest.permission.ACCESS_COARSE_LOCATION}, 100);
         }
     }
 
-    /**
-     * This method get the current location of the user.
-     * @return
-     * Return the geopoint of current location
-     */
-    @Nullable
-    private GeoPoint getLocation() {
-        //      Location location = null;
-        LocationManager locationManager = (LocationManager) getActivity().getApplicationContext().getSystemService(Context.LOCATION_SERVICE);
-        if (locationManager == null) {
-            return null;
-        }
-        Location location = null;
-
-
-        if (ActivityCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
-            locationManager.requestLocationUpdates(locationManager.GPS_PROVIDER, 1000L, 0f, new LocationListener() {
-
-                @Override
-                public void onLocationChanged(@NonNull Location location) {
-                }
-            });
-        }
-        List<String> providers = locationManager.getProviders(true);
-        for (String provider : providers) {
-            @SuppressLint("MissingPermission") Location l = locationManager.getLastKnownLocation(provider);
-            if (l == null){
-                continue;
-            }
-            if (location == null||location.getAccuracy() < l.getAccuracy()){
-                location = l;
-            }
-        }
-        if (location!= null){
-            GeoPoint geoPoint = new GeoPoint(location.getLatitude(),location.getLongitude());
-            return geoPoint;
-        }else{
-            return null;
-        }
-    }
+//    /**
+//     * This method get the current location of the user.
+//     * @return
+//     * Return the geopoint of current location
+//     */
+//    @Nullable
+//    public GeoPoint getLocation() {
+//        //      Location location = null;
+//        LocationManager locationManager = (LocationManager) getActivity().getApplicationContext().getSystemService(Context.LOCATION_SERVICE);
+//        if (locationManager == null) {
+//            return null;
+//        }
+//        Location location = null;
+//
+//
+//        if (ActivityCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED && ActivityCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_COARSE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
+//            locationManager.requestLocationUpdates(locationManager.GPS_PROVIDER, 1000L, 0f, new LocationListener() {
+//
+//                @Override
+//                public void onLocationChanged(@NonNull Location location) {
+//                }
+//            });
+//        }
+//        List<String> providers = locationManager.getProviders(true);
+//        for (String provider : providers) {
+//            @SuppressLint("MissingPermission") Location l = locationManager.getLastKnownLocation(provider);
+//            if (l == null){
+//                continue;
+//            }
+//            if (location == null||location.getAccuracy() < l.getAccuracy()){
+//                location = l;
+//            }
+//        }
+//        if (location!= null){
+//            GeoPoint geoPoint = new GeoPoint(location.getLatitude(),location.getLongitude());
+//            return geoPoint;
+//        }else{
+//            return null;
+//        }
+//    }
 
     public void setOnCameraCloseListener(onCameraClose listener) {
         this.listener = listener;
